mcp-servers/mcp_server/requirements.txt
langchain==1.2.10
langchain-openai==1.1.10
mcp==1.26.0
openpyxl==3.1.5
pandas==3.0.1
pathlib==1.0.1
plotly==6.5.2
pydantic==2.12.5
reportlab==4.4.10
python-docx==1.2.0